requirements.txt
Django==1.8.1
django-appconf==1.0.1
django-cleanup==0.3.1
django-compressor==1.5
django-crispy-forms==1.4.0
django-datetime-widget==0.9.3
django-grappelli==2.6.4
django-secretballot==0.4.0
django-taggit==0.14.0
django-watson==1.1.8
feedparser==5.2.0.post1
httplib2==0.9.1
Markdown==2.6.2
oauth2==1.5.211
oauthlib==0.7.2
Pillow==2.6.2
PyJWT==1.1.0
python-openid==2.2.5
python-ptrace==0.8.1
python-slugify==1.0.2
python-social-auth==0.2.9
pytz==2013b0
requests==2.7.0
requests-oauthlib==0.5.0
simplejson==3.6.5
six==1.9.0
sqlparse==0.1.15
unicode-slugify==0.1.3
Unidecode==0.4.17
